requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.3.0
